refactor all actor method and params filename into <actor-name_params> format
